requirements.txt
pandas == 0.23.4
google-api-python-client == 1.7.11
google-auth == 1.6.3
google-auth-httplib2 ==  0.0.3
google-auth-oauthlib ==  0.4.1
pafy == 0.5.4
python-vlc == 3.0.7110
pyQt5 == 5.13.1